Reimplement ``HashableDict``.
BEFORE: We were using Alex Martelli's excellent implementation from
StackOverflow. Unfortunately, this falls down when attempting to create
a ``HashableDict`` with keys that aren't sortable. We had addressed this
once before, but in a not-very-clean fashion, and the switch to the new
implementation lost it.

AFTER: We have a more robust implementation. Rather than attempting to
sort the dictionary after the fact, we extend
``collections.OrderedDict`` to ensure that it iterates in the same
order; we were only sorting to ensure our comparisons were consistent,
and ``OrderedDict`` gives us that without breaking in the case of
non-comparable types as keys in the same ``dict``. Further, we are once
again turnng the ``dict`` values into ``tuple``s where required (for
instance, because ``list`` values will break ``hash``). HOWEVER, in
contrast to our earlier ``tuple``-izing implementation, we do *NOT*
exclude non-iterable values from our hash and/or equality comparisons.